Split calculators to separate files
