+++ src/BUILD
@@ -195,6 +195,10 @@ cc_library(
         "kfs_frontend/kfs_grpc_inference_service.hpp",
         "kfs_frontend/kfs_utils.cpp",
         "kfs_frontend/kfs_utils.hpp",
+        "mediapipe_internal/modelapiovmsadapter.cc",
+        "mediapipe_internal/modelapiovmsadapter.h",
+        "mediapipe_internal/modelapiovmsadapterwrapper.cc",
+        "mediapipe_internal/modelapiovmsadapterwrapper.h",
         "mediapipe_internal/mediapipefactory.cpp",
         "mediapipe_internal/mediapipefactory.hpp",
         "mediapipe_internal/mediapipegraphdefinition.cpp",
@@ -202,6 +206,8 @@ cc_library(
         "mediapipe_internal/mediapipegraphexecutor.cpp",
         "mediapipe_internal/mediapipegraphexecutor.hpp",
         "mediapipe_internal/ovms_calculator.cc",
+        "mediapipe_internal/modelapiovmsinferencecalculator.cc",
+        "mediapipe_internal/modelapiovmssessioncalculator.cc",
         "metric.cpp",
         "metric.hpp",
         "metric_config.cpp",
