+++ src/mediapipe_internal/modelapiovmsadapterwrapper.h
@@ -1,0 +1,33 @@
+#pragma once
+//*****************************************************************************
+// Copyright 2023 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//*****************************************************************************
+#include <memory>
+
+#include "mediapipe/framework/calculator_framework.h"
+#include "mediapipe/framework/port/canonical_errors.h"
+#include "src/mediapipe_internal/ovmscalculator.pb.h"
+namespace mediapipe {
+     namespace ovms {
+
+class OVMSInferenceAdapter;
+
+struct AdapterWrapper {
+    std::unique_ptr<OVMSInferenceAdapter> adapter;
+    AdapterWrapper(OVMSInferenceAdapter* adapter);
+    ~AdapterWrapper();
+};
+}  // namespace ovms
+}  // namespace mediapipe
