--- src/test/mediapipe/graphaddadapter.pbtxt
@@ -1,43 +1,0 @@
-#
-# Copyright 2023 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-input_stream: "in1"
-input_stream: "in2"
-output_stream: "out"
-node {
-  calculator: "ModelAPICalculator"
-  input_stream: "INPUT1:in1"
-  input_stream: "INPUT2:in2"
-  output_stream: "SUM:out"
-  node_options: {
-        [type.googleapis.com / mediapipe.OVMSCalculatorOptions]: {
-          servable_name: "add"
-          servable_version: "1"
-          tag_to_input_tensor_names {
-            key: "INPUT1"
-            value: "input1"
-          }
-          tag_to_input_tensor_names {
-            key: "INPUT2"
-            value: "input2"
-          }
-          tag_to_output_tensor_names {
-            key: "SUM"
-            value: "sum"
-          }
-          config_path: "/ovms/src/test/mediapipe/config_standard_add.json"
-        }
-  }
-}
